requirements.txt
discord.py==1.7.3
requests==2.22.0
transformers==4.10.2
git+https://github.com/Rapptz/discord.py@rewrite
dnspython==1.16.0
PyNaCl==1.3.0
